requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
dateparser==1.2.2
Django==5.2.7
djangorestframework==3.16.1
idna==3.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.10.23
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0

playwright~=1.55.0